Move PlatformUtils to io.realm.utils
