+++ test/src/androidTest/kotlin/io/realm/InstrumentedTests.kt
@@ -19,6 +19,7 @@ package io.realm
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import io.realm.internal.RealmInitializer
+import io.realm.util.PlatformUtils
 import org.junit.After
 import org.junit.Before
 import org.junit.runner.RunWith
