+++ test/src/androidTest/kotlin/io/realm/LinkTests.kt
@@ -15,6 +15,7 @@
  */
 package io.realm
 
+import io.realm.util.PlatformUtils
 import test.link.Child
 import test.link.Parent
 import kotlin.test.AfterTest
