+++ test/src/androidTest/kotlin/io/realm/NotificationTests.kt
@@ -16,6 +16,7 @@
 
 package io.realm
 
+import io.realm.util.PlatformUtils
 import io.realm.util.RunLoopThread
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.delay
