+++ test/src/androidTest/kotlin/io/realm/util/PlatformUtils.kt
@@ -1,0 +1,17 @@
+package io.realm.util
+
+import java.io.File
+import java.nio.file.Files
+import kotlin.io.path.ExperimentalPathApi
+import kotlin.io.path.absolutePathString
+
+actual object PlatformUtils {
+    @ExperimentalPathApi
+    actual fun createTempDir(): String {
+        return Files.createTempDirectory("android_tests").absolutePathString()
+    }
+
+    actual fun deleteTempDir(path: String) {
+        File(path).deleteRecursively()
+    }
+}
